Library changed vector from double to float

+++ src/main/java/org/rag4j/integrations/openai/OpenAIEmbedder.java
@@ -26,7 +26,7 @@ public class OpenAIEmbedder implements Embedder {
     }
 
     @Override
-    public List<Double> embed(String text) {
+    public List<Float> embed(String text) {
         EmbeddingsOptions embeddingsOptions = new EmbeddingsOptions(Collections.singletonList(text));
 
         Embeddings embeddings = this.client.getEmbeddings(this.model, embeddingsOptions);
